Add test for raw kube reconciler
